fix: protect page
